requirements.txt
Django==5.2.9
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-tenants==3.9.0
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
cryptography==46.0.3
PyJWT==2.10.1
